+++ packages/react-native-sdk/src/constants/A11yLabels.ts
@@ -13,6 +13,8 @@ export enum A11yComponents {
 
 export enum A11yButtons {
   PARTICIPANTS_INFO = 'participants-info-button',
+  EXIT_PARTICIPANTS_INFO = 'exit-participants-info-button',
+  HANG_UP_CALL = 'hang-up-call-button',
 }
 
 export enum A11yValues {