+++ sample-apps/react-native/dogfood/src/constants/A11yLabels.ts
@@ -1,0 +1,3 @@
+export enum A11yButtons {
+  LOG_OUT_AVATAR = 'log-out-avatar-button',
+}