+++ sample-apps/react-native/dogfood/src/contexts/AppContext.tsx
@@ -1,0 +1,24 @@
+import createStoreContext from './createStoreContext';
+
+export type AppMode = 'Meeting' | 'Call' | 'None';
+
+type AppGlobalStore = {
+  username: string;
+  userImageUrl?: string;
+  loopbackMyVideo: boolean;
+  appMode: AppMode;
+};
+
+export const {
+  Provider: AppGlobalContextProvider,
+  useStoreValue: useAppGlobalStoreValue,
+  useStoreSetState: useAppGlobalStoreSetState,
+} = createStoreContext<AppGlobalStore>(
+  {
+    username: '',
+    userImageUrl: '',
+    loopbackMyVideo: false,
+    appMode: 'None',
+  },
+  ['username', 'userImageUrl', 'appMode'],
+);