+++ sample-apps/react/react-video-demo/src/components/Views/MeetingView/MeetingView.tsx
@@ -22,6 +22,8 @@ import MeetingLayout from '../../Layout/MeetingLayout';
 
 import { useWatchChannel } from '../../../hooks/useWatchChannel';
 
+import { DeviceSettingsCaptor } from '../../../utils/useDeviceStorage';
+
 import { useTourContext } from '../../../contexts/TourContext';
 import { usePanelContext } from '../../../contexts/PanelContext';
 import { tour } from '../../../../data/tour';
@@ -108,8 +110,6 @@ export const View: FC<Props & Meeting> = ({
 
   useEffect(() => {
     if (!chatClient || !channelWatched) return;
-
-    console.log({ unread, isChatVisible });
 
     if (isChatVisible && unread !== 0) {
       setUnread(0);
@@ -203,6 +203,7 @@ export const View: FC<Props & Meeting> = ({
         participantsAmount={participants?.length}
         participants={participants}
       />
+      <DeviceSettingsCaptor />
     </MeetingLayout>
   );
 };